requirements.txt
gpiozero
pywhatkit
cv2-python